+++ test/custom-resource-handlers/certificate-signing-request.test.ts
@@ -11,6 +11,9 @@ const eventBase = {
   ResponseURL: 'https://response/url',
   RequestId: '5EF100FB-0075-4716-970B-FBCA05BFE118',
   ResourceProperties: {
+    ServiceToken:             'Service-Token (Would be the function ARN',
+    ResourceVersion:          'The hash of the function code',
+
     DnCommonName:             'Test',
     DnCountry:                'FR',
     DnStateOrProvince:        'TestLand',
@@ -20,6 +23,7 @@ const eventBase = {
     DnEmailAddress:           'test@acme.test',
     KeyUsage:                 'critical,use-the-key',
     ExtendedKeyUsage:         'critical,abuse-the-key',
+    PrivateKeySecretId:       'arn:::private-key-secret',
   },
   ResourceType: 'Custom::Resource::Type',
   StackId: 'StackID-1324597',
@@ -52,8 +56,8 @@ keyUsage             = ${eventBase.ResourceProperties.KeyUsage}
 subjectKeyIdentifier = hash`;
 
 jest.spyOn(fs, 'mkdtemp').mockName('fs.mkdtemp')
-  .mockImplementation(async (_, cb) => cb(undefined, mockTmpDir));
-jest.spyOn(fs, 'readFile').mockName('fs.readFile')
+  .mockImplementation(async (_, cb) => cb(undefined as any, mockTmpDir));
+fs.readFile = jest.fn().mockName('fs.readFile')
   .mockImplementation(async (file, opts, cb) => {
     expect(opts.encoding).toBe('utf8');
     switch (file) {
@@ -64,11 +68,12 @@ jest.spyOn(fs, 'readFile').mockName('fs.readFile')
     default:
       cb(new Error('Unexpected call!'));
     }
-  });
-const mockWriteFile = jest.spyOn(fs, 'writeFile').mockName('fs.writeFile')
-  .mockImplementation((_pth, _data, _opts, cb) => cb());
+  }) as any;
+const mockWriteFile = fs.writeFile = jest.fn().mockName('fs.writeFile')
+  .mockImplementation((_pth, _data, _opts, cb) => cb()) as any;
 const mockSecretsManager = createMockInstance(aws.SecretsManager);
-jest.spyOn(aws, 'SecretsManager').mockImplementation(() => mockSecretsManager);
+aws.SecretsManager = jest.fn().mockName('SecretsManager')
+  .mockImplementation(() => mockSecretsManager) as any;
 mockSecretsManager.getSecretValue = jest.fn().mockName('SecretsManager.getSecretValue')
   .mockImplementation(() => ({ promise: () => Promise.resolve({ SecretString: mockPrivateKey }) })) as any;
 const mockExec = jest.fn().mockName('_exec').mockRejectedValue(new Error('Unexpected call!'));
@@ -77,7 +82,7 @@ jest.mock('../../custom-resource-handlers/src/_rmrf', () => mockRmrf);
 const mockRmrf = jest.fn().mockName('_rmrf')
   .mockResolvedValue(undefined);
 jest.mock('../../custom-resource-handlers/src/_rmrf', () => mockRmrf);
-jest.spyOn(cfn, 'sendResponse').mockName('cfn.sendResponse').mockResolvedValue(undefined);
+jest.spyOn(cfn, 'sendResponse').mockName('cfn.sendResponse').mockResolvedValue(Promise.resolve({}));
 
 beforeEach(() => jest.clearAllMocks());