+++ test/run-test.sh
@@ -9,20 +9,25 @@ cdk_app="test/integ.delivlib.js"
 expected="test/expected.json"
 actual="/tmp/actual.json"
 
-if [ "${1:-}" == "diff" ]; then
-  npx cdk --no-version-reporting -a ${cdk_app} diff
-  exit 0
-fi
+export TEST_STACK_NAME="delivlib-test"
 
 if [ "${1:-}" == "synth" ]; then
   npx cdk --no-version-reporting -a ${cdk_app} synth
   exit 0
 fi
 
-npx cdk --no-version-reporting -a ${cdk_app} synth > ${actual}
+npx cdk --no-version-reporting --no-asset-metadata -a ${cdk_app} synth > ${actual}
 
 if [ "${1:-}" == "update" ]; then
+  hash="$(cat ${actual} | shasum | cut -c1-6 | xargs)"
+  export TEST_STACK_NAME="delivlib-test-${hash}"
   npx cdk --no-version-reporting -a ${cdk_app} deploy ${2:-} ${3:-} ${4:-}
+  echo "Stack deployed, now, go to the console and wait for the pipeline to fully stabalize"
+  echo "Press ENTER to confirm that pipeline is green"
+  read
+  echo "Okay, now go to CFN console and delete the test stack ${TEST_STACK_NAME}"
+  echo "Press ENTER to confirm that the stack has been deleted"
+  read
   cp -f ${actual} ${expected}
 fi